refactor(web): decompose provider CredentialsRotateDialog under ESLint size caps
Extract useRotateState hook, buildRotatePayload helper, and per-auth-type field
sub-components. Pure size refactor. Part of #2066.
