+++ web/src/pages/providers/CredentialsRotateDialog.tsx
@@ -31,14 +31,31 @@ const ROTATABLE_AUTH_TYPES: ReadonlySet<AuthType> = new Set([
  * backend rejects with HTTP 422 if the variant does not match, which
  * the store surfaces as an error toast.
  */
-export function CredentialsRotateDialog({
-  providerName,
-  provider,
-  open,
-  onClose,
-}: CredentialsRotateDialogProps) {
-  const rotateCredentials = useProvidersStore((s) => s.rotateCredentials)
+interface RotateState {
+  apiKey: string
+  setApiKey: (v: string) => void
+  subscriptionToken: string
+  setSubscriptionToken: (v: string) => void
+  tosAccepted: boolean
+  setTosAccepted: (v: boolean) => void
+  headerName: string
+  setHeaderName: (v: string) => void
+  headerValue: string
+  setHeaderValue: (v: string) => void
+  oauthTokenUrl: string
+  setOauthTokenUrl: (v: string) => void
+  oauthClientId: string
+  setOauthClientId: (v: string) => void
+  oauthClientSecret: string
+  setOauthClientSecret: (v: string) => void
+  oauthScope: string
+  setOauthScope: (v: string) => void
+  submitting: boolean
+  setSubmitting: (v: boolean) => void
+  reset: () => void
+}
 
+function useRotateState(): RotateState {
   const [apiKey, setApiKey] = useState('')
   const [subscriptionToken, setSubscriptionToken] = useState('')
   const [tosAccepted, setTosAccepted] = useState(false)
@@ -49,11 +66,6 @@ export function CredentialsRotateDialog({
   const [oauthClientSecret, setOauthClientSecret] = useState('')
   const [oauthScope, setOauthScope] = useState('')
   const [submitting, setSubmitting] = useState(false)
-
-  if (provider === null || providerName === null) return null
-
-  const authType = provider.auth_type
-  const supported = ROTATABLE_AUTH_TYPES.has(authType)
 
   const reset = (): void => {
     setApiKey('')
@@ -68,49 +80,174 @@ export function CredentialsRotateDialog({
     setSubmitting(false)
   }
 
+  return {
+    apiKey, setApiKey, subscriptionToken, setSubscriptionToken, tosAccepted, setTosAccepted,
+    headerName, setHeaderName, headerValue, setHeaderValue, oauthTokenUrl, setOauthTokenUrl,
+    oauthClientId, setOauthClientId, oauthClientSecret, setOauthClientSecret,
+    oauthScope, setOauthScope, submitting, setSubmitting, reset,
+  }
+}
+
+function buildRotatePayload(
+  authType: AuthType,
+  s: RotateState,
+): CredentialsRotateRequest | null {
+  if (authType === 'api_key') return { auth_type: 'api_key', api_key: s.apiKey }
+  if (authType === 'subscription') {
+    return {
+      auth_type: 'subscription',
+      subscription_token: s.subscriptionToken,
+      tos_accepted: s.tosAccepted,
+    }
+  }
+  if (authType === 'custom_header') {
+    return {
+      auth_type: 'custom_header',
+      custom_header_name: s.headerName,
+      custom_header_value: s.headerValue,
+    }
+  }
+  if (authType === 'oauth') {
+    return {
+      auth_type: 'oauth',
+      oauth_token_url: s.oauthTokenUrl,
+      oauth_client_id: s.oauthClientId,
+      oauth_client_secret: s.oauthClientSecret,
+      ...(s.oauthScope.trim() ? { oauth_scope: s.oauthScope.trim() } : {}),
+    }
+  }
+  return null
+}
+
+function ApiKeyAuthField({ s }: { s: RotateState }) {
+  return (
+    <InputField
+      label="New API key"
+      type="password"
+      value={s.apiKey}
+      onChange={(e) => s.setApiKey(e.target.value)}
+      required
+    />
+  )
+}
+
+function SubscriptionAuthFields({ s }: { s: RotateState }) {
+  return (
+    <>
+      <InputField
+        label="New subscription token"
+        type="password"
+        value={s.subscriptionToken}
+        onChange={(e) => s.setSubscriptionToken(e.target.value)}
+        required
+      />
+      <ToggleField
+        label="I accept the subscription Terms of Service"
+        checked={s.tosAccepted}
+        onChange={s.setTosAccepted}
+      />
+    </>
+  )
+}
+
+function CustomHeaderAuthFields({ s }: { s: RotateState }) {
+  return (
+    <>
+      <InputField
+        label="Header name"
+        value={s.headerName}
+        onChange={(e) => s.setHeaderName(e.target.value)}
+        required
+      />
+      <InputField
+        label="Header value"
+        type="password"
+        value={s.headerValue}
+        onChange={(e) => s.setHeaderValue(e.target.value)}
+        required
+      />
+    </>
+  )
+}
+
+function OauthAuthFields({ s }: { s: RotateState }) {
+  return (
+    <>
+      <InputField
+        label="Token URL"
+        value={s.oauthTokenUrl}
+        onChange={(e) => s.setOauthTokenUrl(e.target.value)}
+        required
+      />
+      <InputField
+        label="Client ID"
+        value={s.oauthClientId}
+        onChange={(e) => s.setOauthClientId(e.target.value)}
+        required
+      />
+      <InputField
+        label="Client secret"
+        type="password"
+        value={s.oauthClientSecret}
+        onChange={(e) => s.setOauthClientSecret(e.target.value)}
+        required
+      />
+      <InputField
+        label="Scope"
+        hint="Optional"
+        value={s.oauthScope}
+        onChange={(e) => s.setOauthScope(e.target.value)}
+      />
+    </>
+  )
+}
+
+function RotateAuthFields({ authType, s }: { authType: AuthType; s: RotateState }) {
+  return (
+    <div className="mt-section-gap flex flex-col gap-grid-gap">
+      {authType === 'api_key' && <ApiKeyAuthField s={s} />}
+      {authType === 'subscription' && <SubscriptionAuthFields s={s} />}
+      {authType === 'custom_header' && <CustomHeaderAuthFields s={s} />}
+      {authType === 'oauth' && <OauthAuthFields s={s} />}
+    </div>
+  )
+}
+
+export function CredentialsRotateDialog({
+  providerName,
+  provider,
+  open,
+  onClose,
+}: CredentialsRotateDialogProps) {
+  const rotateCredentials = useProvidersStore((s) => s.rotateCredentials)
+  const state = useRotateState()
+
+  if (provider === null || providerName === null) return null
+
+  const authType = provider.auth_type
+  const supported = ROTATABLE_AUTH_TYPES.has(authType)
+
+  const closeAndReset = (): void => {
+    state.reset()
+    onClose()
+  }
+
   const handleSubmit = async (): Promise<void> => {
     // Guard against duplicate submissions: rapid clicks before the
     // store mutation resolves would issue parallel rotation
     // requests, each writing a separate audit row.
-    if (submitting) return
-    let payload: CredentialsRotateRequest
-    if (authType === 'api_key') {
-      payload = { auth_type: 'api_key', api_key: apiKey }
-    } else if (authType === 'subscription') {
-      payload = {
-        auth_type: 'subscription',
-        subscription_token: subscriptionToken,
-        tos_accepted: tosAccepted,
-      }
-    } else if (authType === 'custom_header') {
-      payload = {
-        auth_type: 'custom_header',
-        custom_header_name: headerName,
-        custom_header_value: headerValue,
-      }
-    } else if (authType === 'oauth') {
-      payload = {
-        auth_type: 'oauth',
-        oauth_token_url: oauthTokenUrl,
-        oauth_client_id: oauthClientId,
-        oauth_client_secret: oauthClientSecret,
-        ...(oauthScope.trim() ? { oauth_scope: oauthScope.trim() } : {}),
-      }
-    } else {
-      return
-    }
-    setSubmitting(true)
+    if (state.submitting) return
+    const payload = buildRotatePayload(authType, state)
+    if (payload === null) return
+    state.setSubmitting(true)
     try {
       const result = await rotateCredentials(providerName, payload)
-      if (result !== null) {
-        reset()
-        onClose()
-      }
+      if (result !== null) closeAndReset()
     } finally {
       // Always clear ``submitting`` even if the store mutation
       // throws past the sentinel contract; otherwise a one-off bug
       // would leave the dialog permanently disabled.
-      setSubmitting(false)
+      state.setSubmitting(false)
     }
   }
 
@@ -118,10 +255,7 @@ export function CredentialsRotateDialog({
     <Dialog.Root
       open={open}
       onOpenChange={(next) => {
-        if (!next) {
-          reset()
-          onClose()
-        }
+        if (!next) closeAndReset()
       }}
     >
       <Dialog.Portal>
@@ -143,99 +277,18 @@ export function CredentialsRotateDialog({
             />
           )}
 
-          {supported && (
-            <div className="mt-section-gap flex flex-col gap-grid-gap">
-              {authType === 'api_key' && (
-                <InputField
-                  label="New API key"
-                  type="password"
-                  value={apiKey}
-                  onChange={(e) => setApiKey(e.target.value)}
-                  required
-                />
-              )}
-              {authType === 'subscription' && (
-                <>
-                  <InputField
-                    label="New subscription token"
-                    type="password"
-                    value={subscriptionToken}
-                    onChange={(e) => setSubscriptionToken(e.target.value)}
-                    required
-                  />
-                  <ToggleField
-                    label="I accept the subscription Terms of Service"
-                    checked={tosAccepted}
-                    onChange={setTosAccepted}
-                  />
-                </>
-              )}
-              {authType === 'custom_header' && (
-                <>
-                  <InputField
-                    label="Header name"
-                    value={headerName}
-                    onChange={(e) => setHeaderName(e.target.value)}
-                    required
-                  />
-                  <InputField
-                    label="Header value"
-                    type="password"
-                    value={headerValue}
-                    onChange={(e) => setHeaderValue(e.target.value)}
-                    required
-                  />
-                </>
-              )}
-              {authType === 'oauth' && (
-                <>
-                  <InputField
-                    label="Token URL"
-                    value={oauthTokenUrl}
-                    onChange={(e) => setOauthTokenUrl(e.target.value)}
-                    required
-                  />
-                  <InputField
-                    label="Client ID"
-                    value={oauthClientId}
-                    onChange={(e) => setOauthClientId(e.target.value)}
-                    required
-                  />
-                  <InputField
-                    label="Client secret"
-                    type="password"
-                    value={oauthClientSecret}
-                    onChange={(e) => setOauthClientSecret(e.target.value)}
-                    required
-                  />
-                  <InputField
-                    label="Scope"
-                    hint="Optional"
-                    value={oauthScope}
-                    onChange={(e) => setOauthScope(e.target.value)}
-                  />
-                </>
-              )}
-            </div>
-          )}
+          {supported && <RotateAuthFields authType={authType} s={state} />}
 
           <div className="mt-section-gap flex justify-end gap-grid-gap">
-            <Button
-              variant="secondary"
-              onClick={() => {
-                reset()
-                onClose()
-              }}
-              disabled={submitting}
-            >
+            <Button variant="secondary" onClick={closeAndReset} disabled={state.submitting}>
               Cancel
             </Button>
             <Button
               variant="destructive"
               onClick={() => void handleSubmit()}
-              disabled={!supported || submitting}
+              disabled={!supported || state.submitting}
             >
-              {submitting ? 'Rotating…' : 'Rotate credentials'}
+              {state.submitting ? 'Rotating…' : 'Rotate credentials'}
             </Button>
           </div>
         </Dialog.Popup>
